chore: fix tests for engineers running on Java 9+
For engineers running on Java 9 or above, some of the tests (like
`SerializerTest`) fail.

This is because when the client receives errors from Weaviate, it
attempts to deserialize into `Exception`s through the use of Gson. Gson
internally uses reflection on internal variables like
`Throwable#detailMessage`, which is barred due to the "strict
encapsulation" module system introduced in Java 9 and above.

The result is an illegal reflection error.

To fix this, the internal `java.lang` module was opened up so tests run
as intended for contributors running Java 9 or above.

As this is the case for end users as well, a message was added to the
README to let users of the library know with instructions to tweak
their build tools to work around this.