feat: implement simple NearVector search
This is a first iteration, so I reused a lot of the 'experimental' code
we wrote to 5.1.0-alpha1 and also did not care to much about code duplication
and project structure.
Will be refactored in the following iterations.

Integration tests now have a set of utils, which allows them to re-use the same
Weaviate instance across multiple test suites without collection name clashes.