refactor: replace Float[] with float[]
We started off using Float[] and Float[][] because v5 did so
and it was easier to reuse (read: copy-paste) some bits of code.

Thing with object arrays is that they name N allocations (1 for each element)
compared to 1 that a primitive array makes.
That creates unnecessary work for GC and we aren't gaining anything from
using boxed values.

They're still useful for request parameters like distance/certainty/force,
because they can be null. But vectors are not sparse and should never contain
nulls in principle.