test: delete server-related assertion
This test simply checks that server returns some string exactly,
which is based _entirely_ on how the server is configured and
has no relation to the client-side logic. It is deserialized in
and Object, so there is no special behavior to verify.

The assertion does fail with v1.33.0-rc.0 though, because
the server started returning a different string.

Once again, this string is not something that affects client
behavior in any way.