feat: read/write data with POJOs
internal.orm package now knows how to marshal
and scan the response back into custom class instances.
To simplify the implementation, we only support 'record'
types for the moment, because they are final classes and
do not have an 'extends' clause.

Array[] and List() containers are supported,
other container types not.